# Improve performance.