Before
```
$ go test -bench=.
BenchmarkCamelToSnake-8   	  100000	     17449 ns/op
PASS
ok  	github.com/fredr/caseconv	1.927s
```
After
```
$ go test -bench=.
BenchmarkCamelToSnake-8   	  300000	      4345 ns/op
PASS
ok  	github.com/fredr/caseconv	1.358s
```